+++ packages/addon-mixins/.npmignore
@@ -1,0 +1,7 @@
+src/
+tests/
+types/
+.eslintrc
+*.log
+*.tsbuildinfo
+tsconfig.json
