--- packages/system/src/mixins/index.ts
@@ -1,10 +1,0 @@
-export * from './background';
-export * from './border';
-export * from './display';
-export * from './foreground';
-export * from './heading';
-export * from './reset';
-export * from './root';
-export * from './shadow';
-export * from './text';
-export * from './ui';
