+++ rollup.config.js
@@ -13,6 +13,7 @@ const packages = [
   'types',
   'utils',
   'addon-direction',
+  'addon-mixins',
   'addon-vendor',
   'system',
   'style',
